Delete ui.js
This file contained all our UI related functions, we are now splitting them into modules accordingly to their responsabilities
